+++ packages/@expo/cli/e2e/__tests__/export/single-page-app.test.ts
@@ -23,6 +23,7 @@ describe('exports with single-page', () => {
           EXPO_USE_STATIC: 'single',
           E2E_ROUTER_SRC: 'static-rendering',
           E2E_ROUTER_ASYNC: 'development',
+          EXPO_USE_FAST_RESOLVER: 'true',
         },
       });
     },