+++ packages/@expo/cli/e2e/__tests__/export/static-rendering.test.ts
@@ -26,6 +26,7 @@ describe('exports static', () => {
             EXPO_USE_STATIC: 'static',
             E2E_ROUTER_SRC: 'static-rendering',
             E2E_ROUTER_ASYNC: 'development',
+            EXPO_USE_FAST_RESOLVER: 'true',
           },
         }
       );